main
The aim of this project is the preliminary design of the fuselage skin of a commercial airliner. The
fuselage is represented by a cylindrical tube of 5.75m diameter under internal pressure, taking into
account the effect of the tail and the nose. We treat the cylindrical skin of the fuselage locally like a
plate under biaxial tension. The elastic properties of a carbon/epoxy ply, of
which the composite is composed and its tensile, compressive, and shear strengths
are given.